+++ com_bears_aichatbot/src/Controller/DisplayController.php
@@ -1,0 +1,15 @@
+<?php
+/**
+ * Display controller for com_bears_aichatbot (administrator)
+ */
+
+namespace Joomla\Component\Bears_aichatbot\Administrator\Controller;
+
+\defined('_JEXEC') or die;
+
+use Joomla\CMS\MVC\Controller\BaseController;
+
+class DisplayController extends BaseController
+{
+    protected $default_view = 'dashboard';
+}